refactor: ViewModel 분리
